Move WeightInfo, Config, Event into Statistics pallet.
