+++ pallets/common/src/traits/mod.rs
@@ -27,4 +27,3 @@ pub mod balances;
 pub mod checkpoint;
 pub mod compliance_manager;
 pub mod settlement;
-pub mod statistics;
