+++ pallets/statistics/src/benchmarking.rs
@@ -4,7 +4,6 @@ use sp_std::collections::btree_set::BTreeSet;
 use sp_std::prelude::*;
 
 use pallet_identity::benchmarking::{User, UserBuilder};
-use polymesh_common_utilities::traits::asset::Config as Asset;
 use polymesh_primitives::bench::create_and_issue_sample_asset;
 use polymesh_primitives::constants::currency::{ONE_UNIT, POLY};
 use polymesh_primitives::{jurisdiction::*, statistics::*, Claim, ClaimType, Scope};
@@ -81,14 +80,14 @@ fn make_transfer_conditions(stats: &BTreeSet<StatType>, count: u32) -> BTreeSet<
         .collect()
 }
 
-fn init_asset<T: Asset>() -> (User<T>, AssetId) {
+fn init_asset<T: Config>() -> (User<T>, AssetId) {
     let owner = UserBuilder::<T>::default().generate_did().build("OWNER");
     let asset_id =
         create_and_issue_sample_asset::<T>(owner.account(), true, None, b"MyAsset", true);
     (owner, asset_id)
 }
 
-fn init_transfer_conditions<T: Config + Asset>(
+fn init_transfer_conditions<T: Config>(
     count_stats: u32,
     count_conditions: u32,
 ) -> (
@@ -103,7 +102,7 @@ fn init_transfer_conditions<T: Config + Asset>(
     (owner, asset_id, stats, conditions)
 }
 
-fn init_exempts<T: Config + Asset>(
+fn init_exempts<T: Config>(
     count: u32,
 ) -> (User<T>, TransferConditionExemptKey, BTreeSet<IdentityId>) {
     let (owner, asset_id) = init_asset::<T>();
@@ -198,8 +197,6 @@ mod limits {
 }
 
 benchmarks! {
-    where_clause { where T: Asset }
-
     set_active_asset_stats {
         let i in 1..T::MaxStatsPerAsset::get().saturating_sub(1);
